+++ src/pages/LoginPage.tsx
@@ -1,0 +1,7 @@
+import Login from '../components/loginpage/Login';
+
+function LoginPage() {
+  return <Login />;
+}
+
+export default LoginPage;
